break vimgolf.py into several files
