--- vimgolf/__main__.py
@@ -1,6 +1,0 @@
-import sys
-
-from vimgolf.vimgolf import main
-
-if __name__ == "__main__":
-    sys.exit(main())
